moved sample and default config to source root
